# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GenAI Email & Report Drafting API
  version: 1.0.0
  description: Flask REST API for AI-powered email and report drafting.
servers:
  - url: /
    description: Same origin as the Swagger UI host
paths:
...
  /api/documents/email:generate:
    post:
      summary: Generate an email draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailGenerateRequest"
      responses:
        "201":
          description: Email generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/report:generate:
    post:
      summary: Generate a report draft
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportGenerateRequest"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/summary:
    get:
      summary: System summary metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users:
    post:
      summary: Create a new admin user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Admin user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/audit-logs:
    get:
      summary: Get audit logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Audit log list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: string
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        created_at:
          type: string
          format: date-time
    EmailGenerateRequest:
      type: object
      required: [context]
      properties:
        context:
          type: string
          maxLength: 5000
        recipient:
          type: string
          maxLength: 500
        subject:
          type: string
          maxLength: 500
        tone:
          type: string
          enum: [professional, casual, formal, friendly]
    ReportGenerateRequest:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          maxLength: 5000
        key_points:
          type: string
          maxLength: 5000
        tone:
          type: string
          enum: [professional, casual, formal, friendly]
        structure:
          type: string
          enum: [executive_summary, detailed, bullet_points]
    Document:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        doc_type:
          type: string
          enum: [email, report]
        title:
          type: string
          nullable: true
        prompt_input:
          type: string
        content:
          type: string
        tone:
          type: string
        structure:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    DocumentResponse:
      type: object
      properties:
        message:
          type: string
        document:
          $ref: "#/components/schemas/Document"
        request_id:
          type: string
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        action:
          type: string
        entity_type:
          type: string
          nullable: true
        entity_id:
          type: integer
          nullable: true
        request_context_id:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
    AuditLogResponse:
      type: object
      properties:
        audit_logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AdminSummary:
      type: object
      properties:
        total_users:
          type: integer
        total_documents:
          type: integer
        documents_last_24h:
          type: integer
        recent_events_count:
          type: integer